Time: 0 ms (100%), Space: 41.4 MB (37.55%) - LeetHub

+++ 1411-convert-binary-number-in-a-linked-list-to-integer/1411-convert-binary-number-in-a-linked-list-to-integer.java
@@ -15,6 +15,6 @@ class Solution {
             sb.append(head.val);
             head = head.next;
         }
-        return (int)Long.parseLong(sb.toString(), 2);
+        return Integer.parseInt(sb.toString(), 2);
     }
 }
